Yeni auditlogs yapısı kuruldu #93

+++ api/utils/crudHelper.js
@@ -5,20 +5,27 @@ const fs = require("fs");
 const path = require("path");
 const config = require("../config");
 const { v4: uuidv4 } = require("uuid");
+const AuditLog = require("../db/models/AuditLogs");
 
-const createSlug = (name) => {
-  return name.toLowerCase().replace(/ /g, '-') + '-' + uuidv4(); // Kategori adını slug'a çevir ve uuid ekle
+const createSlug = name => {
+  return name.toLowerCase().replace(/ /g, "-") + "-" + uuidv4(); // Kategori adını slug'a çevir ve uuid ekle
 };
 
 const createEntity = async (Model, body, file, user) => {
   try {
     if (!body.name) {
-      throw new CustomError(Enum.HTTP_CODES.BAD_REQUEST, i18next.t("Category name is required."));
+      throw new CustomError(
+        Enum.HTTP_CODES.BAD_REQUEST,
+        "Category name is required."
+      );
     }
 
     const existingCategory = await Model.findOne({ name: body.name });
     if (existingCategory) {
-      throw new CustomError(Enum.HTTP_CODES.CONFLICT, i18next.t("Category already exists."));
+      throw new CustomError(
+        Enum.HTTP_CODES.CONFLICT,
+        "Category already exists."
+      );
     }
 
     // Slug oluşturma
@@ -28,7 +35,7 @@ const createEntity = async (Model, body, file, user) => {
       name: body.name,
       slug: slug,
       is_active: body.is_active,
-      created_by: user._id,
+      created_by: user.id,
       image: file ? file.filename : undefined,
       tags: body.tags ? JSON.parse(body.tags) : [],
       description: body.description || "",
@@ -36,19 +43,47 @@ const createEntity = async (Model, body, file, user) => {
     });
 
     await newEntity.save();
+
+    // Audit log kaydı oluştur
+    const auditLog = new AuditLog({
+      user_id: user.id,
+      action: "create", // 'create' olarak güncellendi
+      entity: Model.collection.collectionName,
+      entity_id: newEntity._id, // Doğru şekilde yeni kaydın ID'sini kullan
+      old_value: null, // Yeni oluşturulmuş bir kayıt için eski değer yok
+      new_value: newEntity // Yeni değer
+    });
+
+    try {
+      await auditLog.save();
+    } catch (auditError) {
+      console.error("Audit log kaydedilemedi:", auditError);
+    }
+
     return Response.successResponse(newEntity);
   } catch (error) {
     throw error;
   }
 };
 
-const updateEntity = async (Model, id, body, file) => {
+const updateEntity = async (Model, id, body, file, user) => {
   try {
     const entity = await Model.findById(id);
     if (!entity) {
-      throw new CustomError(Enum.HTTP_CODES.NOT_FOUND, i18next.t("Entity not found."));
+      throw new CustomError(Enum.HTTP_CODES.NOT_FOUND, "Entity not found.");
     }
 
+    // Eski değeri kaydet
+    const oldValue = {
+      name: entity.name,
+      is_active: entity.is_active,
+      tags: entity.tags,
+      description: entity.description,
+      parent_id: entity.parent_id,
+      image: entity.image
+    };
+
+    // Güncelleme işlemleri
     entity.name = body.name || entity.name;
     entity.is_active =
       body.is_active !== undefined ? body.is_active : entity.is_active;
@@ -73,15 +108,43 @@ const updateEntity = async (Model, id, body, file) => {
     }
 
     await entity.save();
+
+    // Audit log kaydı oluştur
+    const auditLog = new AuditLog({
+      user_id: user.id, // Kullanıcının ID'sini al
+      action: "update", // Güncelleme işlemi
+      entity: Model.collection.collectionName,
+      entity_id: entity._id,
+      old_value: oldValue, // Eski değer
+      new_value: entity // Yeni değer
+    });
+
+    await auditLog.save();
+
     return Response.successResponse(entity);
   } catch (error) {
     throw error;
   }
 };
 
-const deleteEntities = async (Model, ids, fileDelete = true) => {
+const deleteEntities = async (Model, ids, user, fileDelete = true) => {
   try {
     const entitiesToDelete = await Model.find({ _id: { $in: ids } });
+
+    if (entitiesToDelete.length === 0) {
+      throw new CustomError(Enum.HTTP_CODES.NOT_FOUND, "No entities found.");
+    }
+
+    // Silinecek kayıtların eski değerlerini kaydet
+    const oldValues = entitiesToDelete.map(entity => ({
+      _id: entity._id,
+      name: entity.name,
+      is_active: entity.is_active,
+      tags: entity.tags,
+      description: entity.description,
+      parent_id: entity.parent_id,
+      image: entity.image
+    }));
 
     await Model.deleteMany({ _id: { $in: ids } });
 
@@ -96,30 +159,74 @@ const deleteEntities = async (Model, ids, fileDelete = true) => {
       }
     }
 
+    // Audit log kaydı oluştur
+    for (const id of ids) {
+      const auditLog = new AuditLog({
+        user_id: user.id,
+        action: "delete", // Kalıcı silme işlemi
+        entity: Model.collection.collectionName,
+        entity_id: id, // Her bir silinen kaydın ID'si
+        old_value: entitiesToDelete.find(entity => entity._id.equals(id)), // Eski değer
+        new_value: null // Kalıcı silme sonrası değer yok
+      });
+
+      await auditLog.save();
+    }
+
     return Response.successResponse({ success: true });
   } catch (error) {
     throw error;
   }
 };
 
-const softDeleteEntities = async (Model, ids) => {
+const softDeleteEntities = async (Model, ids, user) => {
   try {
     await Model.updateMany(
       { _id: { $in: ids } },
       { $set: { deleted_at: new Date() } }
     );
+
+    // Audit log kaydı oluştur
+    for (const id of ids) {
+      const auditLog = new AuditLog({
+        user_id: user.id,
+        action: "soft_delete", // Soft delete işlemi
+        entity: Model.collection.collectionName,
+        entity_id: id, // Her bir silinen kaydın ID'si
+        old_value: { deleted_at: null }, // Soft delete öncesi değer
+        new_value: { deleted_at: new Date() } // Soft delete sonrası değer
+      });
+
+      await auditLog.save();
+    }
+
     return Response.successResponse({ success: true });
   } catch (error) {
     throw error;
   }
 };
 
-const restoreEntities = async (Model, ids) => {
+const restoreEntities = async (Model, ids, user) => {
   try {
     await Model.updateMany(
       { _id: { $in: ids } },
       { $set: { deleted_at: null } }
     );
+
+    // Audit log kaydı oluştur
+    for (const id of ids) {
+      const auditLog = new AuditLog({
+        user_id: user.id,
+        action: "restore", // Geri yükleme işlemi
+        entity: Model.collection.collectionName,
+        entity_id: id, // Her bir geri yüklenen kaydın ID'si
+        old_value: { deleted_at: new Date() }, // Geri yüklemeden önceki değer
+        new_value: { deleted_at: null } // Geri yüklemeden sonraki değer
+      });
+
+      await auditLog.save();
+    }
+
     return Response.successResponse({ success: true });
   } catch (error) {
     throw error;
